refactor to core and bo hoc
