+++ src/Text/index.js
@@ -1,23 +1,1 @@
-import React from 'react';
-import {bool} from 'prop-types';
-import {createHOC} from 'wix-ui-core/dist/src/createHOC';
-import Text from './Text';
-import {withEllipsedTooltip} from '../common/EllipsedTooltip';
-
-const EllipsedText = withEllipsedTooltip({showTooltip: true})(Text);
-
-const ProxyText = ({ellipsis, ...props}) => ellipsis ?
-  <EllipsedText {...props}/> :
-  <Text {...props}/>;
-
-ProxyText.propTypes = {
-  ...Text.propTypes,
-
-  /** should the text get ellipsed with tooltip, or should it get broken into lines when it reaches the end of its container */
-  ellipsis: bool
-};
-
-ProxyText.displayName = 'Text';
-
-export default createHOC(ProxyText);
-
+export {default} from './ProxyText';
