+++ src/common/EllipsedTooltip/core/EllipsedTooltip.st.css
@@ -1,0 +1,6 @@
+.root {
+  display: inline-block;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  width: 100%;
+}
